+++ R/find_object_by_value.R
@@ -6,20 +6,24 @@
 #' with an environment (e.g. a \code{\link[base:function]{function}} and a
 #' \code{\link[base:tilde]{formula}}), to start search from.
 #'
-#' @param first If TRUE, the first occurance of `object` among the parent
-#' frames is identified, otherwise the last.
+#' @param which If `"first"` or `"last"`, then the first or the last
+#' occurance of `object` among the parent frames is identified and returned.
+#' If `"all"`, then all occurances are returned.
 #'
 #' @return
-#' A named list with elements `name` and `envir`, where `name` is the
-#' name of `object` as it is named in environment `envir`, i.e.
+#' If `which = "first"` or `which = "last"`, then a named list with
+#' elements `name` and `envir`, where `name` is the name of `object`
+#' as it is named in environment `envir`, i.e.
 #' `identical(envir[[name]], object)`.
+#' If `which = "all"`, then a list of (name, environment) lists are
+#' returned; one for each matching occurence.
 #' If the object could not be located when searching from environment
 #' `from`, then NULL is returned.
 #'
 #' @example incl/find_object_by_value_1.R
 #'
 #' @export
-find_object_by_value <- function(object, from = parent.frame(), first = TRUE) {
+find_object_by_value <- function(object, from = parent.frame(), which = c("first", "last", "all")) {
   if (inherits(from, "environment")) {
     envir <- from
   } else {
@@ -28,7 +32,7 @@ find_object_by_value <- function(object, from = parent.frame(), first = TRUE) {
       stop(sprintf("Argument 'from' does not specify an environment or an object with an environment: %s", mode(from)))
     }
   }
-  stopifnot(length(first) == 1L, is.logical(first), !is.na(first))
+  which <- match.arg(which)
   
   mode <- mode(object) ## scan for objects of this mode
 
@@ -42,7 +46,7 @@ find_object_by_value <- function(object, from = parent.frame(), first = TRUE) {
     skip <- as.environment("CheckExEnv")
   }
 
-  res <- NULL
+  res <- list()
   while (!identical(envir, emptyenv())) {
     ## Skip?
     if (identical(envir, skip)) {
@@ -61,14 +65,17 @@ find_object_by_value <- function(object, from = parent.frame(), first = TRUE) {
         tmp <- get(name, mode = mode, envir = envir, inherits = FALSE)
         ## ... are the identical?
         if (identical(tmp, object, ignore.environment = FALSE, ignore.bytecode = FALSE, ignore.srcref = FALSE)) {
-          res <- list(name = name, envir = envir)
-          if (first) return(res)
+          res_name <- list(name = name, envir = envir)
+          if (which == "first") return(res_name)
+          res <- c(res, list(res_name))
         }
       }
     }
     envir <- parent.env(envir)
   }
 
+  if (length(res) == 0L) return(NULL)
+  if (which == "last") res <- res[[length(res)]]
   res
 }